+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [Unreleased]
+
+### Added
+
+- Add preview support for DFR3Mapping, Dataset, and Hazard when selecting resources from services [#180](https://github.com/IN-CORE/incore-studio/issues/180)
 
 ### Fixed